Some schools don't remove appointments if you are not required to be present (opendag par example). This feature will allow you to remove appointments if certain criteria are met.
In the config you can tell when we should act (I don't think you'll want to cancel the appointment if you forgot your books though), change the values by simply adding or removing a '_' (underscore disbles it). Do note that if you are not permitted to skip the class (a.k.a. you just didn't show up) the appointment won't be removed.